Add fix from https://github.com/lief-project/LIEF/issues/404
This will allow us to remove the dropping of the GIL when loading
extension modules on Windows, should we decide we should do that.

From my POV I am not sure that we should because I do not believe
that LIEF will be the only project ambushed by this change of
behaviour.